refactor: Reorganize member service
